requirements-dev.txt
pytest==9.0.2
pytest-django==4.12.0
flake8==7.3.0
isort==8.0.0
